Refactor PointChart component and update import path
